ModelResource doesn't have a root layer
Rather than making ModelResource specify a root layer -- which should always be a GroupLayer, and which all Models should have -- it instead specifies all the children that sit below root. ModelResource.build() creates the root GroupLayer.

This shortens our ModelResource JSON a bit, and it also means that ModelAnims can no longer be applied to root, which is appropriate: a ModelAnim is for controlling the bits *inside* a Model. Client code can manipulate the Model's root layer without having to worry about a running anim undoing its changes.